new feature

+++ bin/beat.sh
@@ -1,12 +1,21 @@
 #!/bin/bash
 
+
 pid=`ps -ef| grep "celery beat" | grep -v grep | head -n 1 | awk '{print $2}'`
 
-
-while(($pid>0))
+while true
 do
-    kill -9 $pid
-    pid=`ps -ef| grep "celery beat" | grep -v grep | head -n 1 | awk '{print $2}'`
+    if [ ! -n "$pid" ]; then
+        echo "null"
+        break
+    else
+        echo "not null"
+        kill -9 $pid
+        pid=`ps -ef| grep "celery beat" | grep -v grep | head -n 1 | awk '{print $2}'`
+    fi
+    echo PID $pid
+
+    break
 done
 
 
@@ -17,3 +26,4 @@ if [ -a celerybeat.pid ]; then
     rm -rf celerybeat.pid
 fi
 celery beat -A worker.celery
+
